Refactored Routes into individual classes and started configurable message routing.
