+++ hnsecure/src/main/resources/application.properties
@@ -25,6 +25,13 @@ hibc-e45-endpoint="E45"
 hibc-r15-endpoint="R15"
 hibc-r50-endpoint="R50"
 
+e45.protocol=HTTP
+r15.protocol=MQ
+r50.protocol=MQ
+
+# HIBC endpoint
+hibc.uri=http://hibc.hlth.gov.bc.ca
+
 # PharmaNet Endpoint
 pharmanet.uri=https://d1hni-connect.maximusbc.ca/pnetIntegration/submitTransaction
 pharmanet.cert=keystore/CGI-HNI-DEV.pfx
