+++ hnsecure/src/test/resources/application.properties
@@ -20,6 +20,9 @@ file-drops-location = ./
 # Keycloak endpoints
 certs-endpoint = https://common-logon-dev.hlth.gov.bc.ca/auth/realms/v2_pos/protocol/openid-connect/certs
 
+# HIBC endpoint
+hibc.uri=http://hibc.hlth.gov.bc.ca
+
 # PharmaNet Endpoint
 pharmanet.uri=https://d1hni-connect.maximusbc.ca/pnetIntegration/submitTransaction
 pharmanet.cert=keystore/CGI-HNI-DEV.pfx
